+++ src/submodules/pathfinding/Pathfinding.jl
@@ -2,6 +2,7 @@ export Pathfinding
 
 """
     Pathfinding
+
 Submodule containing functionality for path-finding based on the A* algorithm.
 Currently available for [`GridSpace`](@ref) and [`ContinuousSpace`](@ref).
 Discretization of [`ContinuousSpace`](@ref) is taken care of internally.
@@ -10,9 +11,9 @@ You can enable path-finding and set its options by creating an instance of a
 [`Pathfinding.AStar`](@ref) struct. This must be passed to the relevant pathfinding functions
 during the simulation. Call [`plan_route!`](@ref) to set the destination for an agent. This
 triggers the algorithm to calculate a path from the agent's current position to the one
-specified. You can alternatively use [`plan_best_route!`](@ref) to choose the best target
+specified. You can alternatively use [`Pathfinding.plan_best_route!`](@ref) to choose the best target
 from a list. Once a target has been set, you can move an agent one step along its
-precalculated path using the [`move_along_route!`](@ref) function.
+precalculated path using the [`Pathfinding.move_along_route!`](@ref) function.
 
 Refer to the [Maze Solver](https://juliadynamics.github.io/AgentsExampleZoo.jl/dev/examples/maze/),
 [Mountain Runners](https://juliadynamics.github.io/AgentsExampleZoo.jl/dev/examples/runners/)
@@ -41,11 +42,9 @@ module Pathfinding
         delta_cost,
         penaltymap,
         nearby_walkable,
-        random_walkable
-
-    # Deprecations
-    @deprecate set_target! plan_route!
-    @deprecate set_best_target! plan_best_route!
-    @deprecate kill_agent! remove_agent!
+        random_walkable,
+        plan_route!,
+        plan_best_route!,
+        move_along_route!
 
 end # module